Use adapter to manage user registry versions
